requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-socketio==5.10.0
gevent-websocket==0.10.1
psutil==5.9.6
gevent==23.9.1
greenlet==3.0.1
bidict==0.22.1
Werkzeug==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
zope.event==5.0
zope.interface==6.1
requests==2.31.0
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
python-dotenv==1.0.0